Moved logic off ImagePack model and into ImagePacksService
Moved Image Pack menu logic into main menu partial to centralize all the menu code.
