+++ src/partials/Program.Menus.cs
@@ -410,6 +410,60 @@ internal partial class Program
         GlobalHelper.SettingsManager.SaveSettings();
     }
 
+    private static async Task ImagePackSelector(string path)
+    {
+        Console.Clear();
+
+        if (GlobalHelper.PlatformImagePacks.Length > 0)
+        {
+            int choice = 0;
+            var menu = new ConsoleMenu()
+                .Configure(config =>
+                {
+                    config.Selector = "=>";
+                    config.EnableWriteTitle = false;
+                    config.WriteHeaderAction = () => Console.WriteLine("So, what'll it be?:");
+                    config.SelectedItemBackgroundColor = Console.ForegroundColor;
+                    config.SelectedItemForegroundColor = Console.BackgroundColor;
+                });
+
+            foreach (var pack in GlobalHelper.PlatformImagePacks)
+            {
+                menu.Add($"{pack.owner}: {pack.repository} {pack.variant ?? string.Empty}", thisMenu =>
+                {
+                    choice = thisMenu.CurrentItem.Index;
+                    thisMenu.CloseMenu();
+                });
+            }
+
+            menu.Add("Go Back", thisMenu =>
+            {
+                choice = GlobalHelper.PlatformImagePacks.Length;
+                thisMenu.CloseMenu();
+            });
+
+            await menu.ShowAsync();
+
+            if (choice < GlobalHelper.PlatformImagePacks.Length && choice >= 0)
+            {
+                await GlobalHelper.PlatformImagePacksService.Install(path, GlobalHelper.PlatformImagePacks[choice].owner,
+                    GlobalHelper.PlatformImagePacks[choice].repository, GlobalHelper.PlatformImagePacks[choice].variant);
+            }
+            else if (choice == GlobalHelper.PlatformImagePacks.Length)
+            {
+                // What causes this?
+            }
+            else
+            {
+                Console.WriteLine("You fucked up.");
+            }
+        }
+        else
+        {
+            Console.WriteLine("None found. Have a nice day.");
+        }
+    }
+
     private static void SettingsMenu()
     {
         Console.Clear();
